+++ transit_odp/api/urls.py
@@ -12,15 +12,22 @@ from transit_odp.api.views import (
     FaresOpenApiView,
     TimetablesApiView,
     TimetablesViewSet,
+    v2,
 )
 
 app_name = "api"
 
-# TODO - refactor into api_v1
-router = DefaultRouter()
-router.register(r"dataset", TimetablesViewSet, "feed")
-# in a v2 version we should move to timetables/dataset
-router.register(r"fares/dataset", FaresDatasetViewset, "fares-api")
+# TODO - refactor into api_v1,
+router_v1 = DefaultRouter()
+router_v1.register(r"dataset", TimetablesViewSet, "feed")
+router_v1.register(r"fares/dataset", FaresDatasetViewset, "fares-api")
+
+router_v2 = DefaultRouter()
+router_v2.register(r"operators", viewset=v2.OperatorViewSet, basename="operators")
+router_v2.register(r"timetables", viewset=v2.TimetableViewSet, basename="timetables")
+router_v2.register(r"txcfiles", viewset=v2.TimetableFilesViewSet, basename="txcfiles")
+router_v2.register(r"avlfeeds", viewset=v2.DatafeedViewSet, basename="avlfeeds")
+
 schema_view = get_swagger_view(title="Timetables Data API")
 avl_views = get_swagger_view(title="Bus Location Data API")
 fares_views = get_swagger_view(title="Fares Data API")
@@ -63,5 +70,6 @@ else:
 
 urlpatterns += [
     path("app/", include("transit_odp.api.app.urls")),
-    path("v1/", include(router.urls)),
+    path("v1/", include(router_v1.urls)),
+    path("v2/", include((router_v2.urls, app_name), namespace="v2")),
 ]