+++ transit_odp/data_quality/dataclasses/warnings/operators.py
@@ -1,5 +1,0 @@
-from .base import BaseWarning
-
-
-class OperatorMissingSourceNOC(BaseWarning):
-    pass