+++ transit_odp/data_quality/tables/missing_headsign.py
@@ -1,8 +1,0 @@
-from transit_odp.data_quality.tables.base import VehicleJourneyTimingPatternTable
-
-
-class MissingHeadsignWarningTimingTable(VehicleJourneyTimingPatternTable):
-    caption_end = "is missing a destination"
-
-    class Meta(VehicleJourneyTimingPatternTable.Meta):
-        pass