+++ transit_odp/fares/views/dataset_download.py
@@ -1,6 +1,0 @@
-from transit_odp.organisation.constants import DatasetType
-from transit_odp.publish.views.timetable.download import DatasetDownloadView
-
-
-class DownloadFaresFileView(DatasetDownloadView):
-    dataset_type = DatasetType.FARES.value