+++ transit_odp/naptan/naptan_exception.py
@@ -1,2 +1,0 @@
-class NaptanException(Exception):
-    pass